# openapi.yaml
openapi: 3.1.0
info:
  title: Blaxx Pontos API
  version: "0.5.0"
  description: |
    API REST do programa de pontos Blaxx. Compra via PIX, transferencia P2P,
    resgate via PIX, parceiros, beneficios, vouchers, campanhas, notificacoes.

    ## Auth
    Bearer JWT. `POST /auth/login` retorna `access_token` + `refresh_token`.
    Access expira em 30min (S1), refresh em 30 dias.

    ## LGPD
    `DELETE /auth/account` (esquecimento) e `GET /auth/account/export`
    (portabilidade) — art. 18 da Lei 13.709/18.

    ## Versionamento de termos
    `GET /auth/terms/current` devolve versao atual. Se diferente do
    `user.terms_accepted_version`, redirecionar pra re-aceite via
    `POST /auth/terms/reaccept`.
  contact:
    name: Blaxx Pontos
    url: https://blaxxpontos.netlify.app
  license:
    name: Proprietary
servers:
  - url: https://blaxx-pontos-backend.onrender.com
    description: Producao (Render)
  - url: http://localhost:8080
    description: Dev local
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    PixCharge:
      type: object
      properties:
        id: { type: string }
        status: { type: string, enum: [pending, pending_confirmation, paid, expired, refunded, rejected] }
        amount_cents: { type: integer }
        amount_brl: { type: number }
        points_to_credit: { type: integer }
        br_code: { type: string, description: "EMV BR Code" }
        qr_code_image: { type: string, nullable: true, description: "Data-URL PNG" }
        expires_at: { type: string, format: date-time }
paths:
...
  /pix/charge:
    post:
      summary: Cria charge PIX (pacote OU valor livre)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                package:    { type: string, enum: [start, plus, prime, black] }
                amount_brl: { type: number, minimum: 10 }
      responses:
        "201": { description: charge, content: { application/json: { schema: { $ref: "#/components/schemas/PixCharge" } } } }
        "503": { description: "S3-3 BR Code placeholder em prod (BRCODE_NOT_CONFIGURED)" }
  /pix/charge/{charge_id}/events:
    get:
      summary: SSE de status da charge (S4-6)
      parameters:
        - in: path
          name: charge_id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: text/event-stream
          content:
            text/event-stream:
              schema:
                type: string
                example: |
                  event: status
                  data: {"status":"paid","charge_id":"abc","amount_brl":50.0}
...